fix: run placements backfill in batch job mode
PlacementsBackfillService was registered as a hosted service but batch
job mode returns before app.RunAsync(), so the backfill never executed.
Extract the logic into PlacementsBackfiller so it can be called directly
from the batch path.
